backend/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
django-shell-plus==1.1.7
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
future==0.18.2
kombu==4.6.11
psycopg2-binary
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
vine==1.3.0
